+++ CHANGELOG.md
@@ -5,9 +5,13 @@ Version changes are pinned to SDK releases.
 
 ## [Unreleased]
 
+<<<<<<< HEAD
+- program: Refactor fee collection methodology and add some associated instructions. ([#136](https://github.com/zetamarkets/sdk/pull/136))
+=======
 - events: Add new OrderCompleteEvent. ([#135](https://github.com/zetamarkets/sdk/pull/135))
 - events: Remove CancelEvent. ([#135](https://github.com/zetamarkets/sdk/pull/135))
 
+>>>>>>> main
 ## [0.16.3]
 
 - referrals: Fix referral alias fetching bug. ([#139](https://github.com/zetamarkets/sdk/pull/139))